Refactoring: Call newrelic by default first. Log an error if config was already used. Log an error if newrelic is used before initialized
