+++ src/initializers/diamorphosis.ts
@@ -5,7 +5,6 @@ export default (config, orkaOptions: Partial<OrkaOptions>) => {
   config.nodeEnv = config.nodeEnv || 'development';
   config.honeybadgerApiKey = config.honeybadgerApiKey || '';
   config.honeybadgerEnvironment = config.honeybadgerEnvironment || config.nodeEnv;
-  config.newRelicLicenseKey = '';
   config.printLogo = config.printLogo || true;
   config.log = {
     pattern: '%[[%d] [%p] %c%] %m',
